requirements.txt
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.2.1
requests==2.32.5
supabase==2.25.0

